Possibly fix the issue with buttons not being in the center.
I forgot that the margin of 5 on the buttons resulted in 10 XAML units between each tile, and the content is now justified to center. Still haven't fixed the issue with tiles not going next to each other. Maybe I have to shrink larger tiles and use an actual grid with larger tiles having more columnspan and rowspan. Preferably I would have a way to use full-size tiles, but it may take time to figure that out while having four small tiles in a square area like on WP.