Allow the user to press Escape to go back to the top...
of the Start screen and return to the tiles from the All Apps list. This should just work if I can figure out how to send the Escape key to every application as if I'm using a keyboard. I'm going to assume that if another app is currently open, going back to that app will take priority whenever I get around to implementing the window manager. I commented out the keyboard shortcut in my local copy of the code that does the back button to go back to the debugger UI thing, so I think that'll take priority or at least it should also happen.