Add Open Sans files.
I deleted all but Light, Regular, and SemiBold since I don't think I'll need the rest. If I do, I'll add them back in.

Sadly there's no SemiLight like Segoe UI has, but the rest should be close enough for most people. Actually, I just noticed that the header text font is Segoe WP Semilight, so uh, here's hoping nobody notices when I just use Open Sans Light for headers (may just use Regular, not sure). I may need to either file a feature request for Open Sans Semilight or create it myself. I think I'll have to modify Open Sans anyway, as the tail of the Q is slightly longer and in the center more instead of toward the right, and the J goes below the other characters like p, unlike in Segoe UI where it stays on the line with most of the rest of the characters.

I've added most of this message to a reply in #15.

I also moved the OFL license to the wp-metro font's subfolder so it's clear what its license is.