It's finally working as expected, so that the tiles list matches...
the getTilesList() function's output when we tell it not to include the Tile App Name Area Text in its output. We're using its output minus that text to compare to the current tiles list, just so we don't have to write the config file again if nothing changed to prevent extra wear on the eMMC. It's also nice to know that the output is the same. Only issue is the DotDesktopFilePath isn't quoted, even though the tile color is. One thing I had to do was tell Python that we just want the tile widths and heights as integers, since QML doesn't do that for us.