Copy over some files from start to use for the action center.
Deleted some of them since I probably only need a few, like the main window file, pyyaml for loading the Action Center configuration and theme and accent colors, and the main.py file to run things. It won't completely work in this state, but I fixed the font file paths.