Add example toggle for changing the Accent color at runtime...
between two different colors. Here, I'm using Cobalt (#0050ef) as the default and changing it to Maroon (the color I decided to settle on to do color testing back when I first started working on my UXL Launcher ThemeEngine because it's not a commonly-used color) whenever the "CHANGE ACCENT" toggle is toggled on.

Also added a RowLayout so that adding and rearranging buttons is easy. I'm also adding a placeholder property for switching the text to all lowercase for accessibility to help screen readers and anyone that doesn't like stuff to be in all caps because it feels like it's yelling at them.