Switch to using the accent color for tiles and All Apps list items.
I've also removed the option to load Accent colors from the startlayout.yaml file because it was difficult to figure out how to check if TileColor keys were in that file and use a placeholder if they weren't available.

I also think it would end up being a problem if people pin a tile with one Accent color and later decide to change their Accent color, the way I had been handling it, because different tiles would end up with different backgrounds depending on which Accent color they were using at the time.

Another problem I thought of was that a primary tile could end up with a background different from the app in the All Apps list, potentially confusing people.

My main reasoning for removing this feature aside from what I wrote above is that it doesn't really make sense to include non-layout stuff like Accent colors in a file meant to store the layout for tiles. I will, however, try to figure out how to get Accent colors from .desktop files, and those can in turn have their Accent colors overridden if they're copied to a subfolder in the Home directory. That's not implemented yet, though.

The StartLayoutSchemaVersion has been incremented to 0.2 due to the loss of the TileColor property. TileColor will be ignored if it's still in your config files, so you don't have to do anything, but it won't be removed automatically.