Allowing the Back button to be tapped to exit multitasking.
We also now allow leaving multitasking with the "Escape" key, and that's only active when in multitasking. Upon activation, we leave multitasking. The shortcut is automatically set to whether the grid is in overview or not, so we don't have to turn it on or off.

We also ensure input events don't get sent to the windows displayed in multitasking, otherwise we could accidentally have them register the Back button.

I also changed the window size to match the rest of the Retiled apps, which also makes it easier to view the compositor at 1x scaling (like on regular 1080p desktop monitors at 100% scaling).

The Back button is also set to not be focused, which should help with keyboard usage. Might be a good idea to not focus the rest too, so I did.

Still having trouble with popup windows and message boxes, where they close even though I'm pretty sure they aren't supposed to. A possible solution for now would be to send a D-Bus message that says to force sending "Escape", as well as keeping track of the number of message boxes. (see this comment for more details: https://github.com/DrewNaylor/Retiled/issues/133#issuecomment-1289266037 )